+++ software/Core/Inc/set_point.h
@@ -8,24 +8,24 @@
 #include "stdint.h"
 
 typedef struct{
-    float min_stick_output;
-    float max_stick_output;
+    int16_t min_stick_output;
+    int16_t max_stick_output;
 }stick_limits_t;
 
 typedef struct {
-    float min_imu_output;
-    float max_imu_output;
+    int16_t min_imu_output;
+    int16_t max_imu_output;
 }imu_limits_t;
 
 typedef struct {
-    float stick_output;
+    int16_t stick_output;
     stick_limits_t stick_limits;
     imu_limits_t imu_limits;
-    float  set_point;
-    float imu_output;
+    int16_t set_point;
+    int16_t imu_output;
 } set_point_t;
 
-void limits_init(set_point_t *set_point, float min_imu_output, float max_imu_output);
-void set_point_calculation(set_point_t *set_point, float stick_output);
+void limits_init(set_point_t *set_point, int16_t min_imu_output, int16_t max_imu_output);
+void set_point_calculation(set_point_t *set_point, int16_t stick_output);
 
 #endif //BETTERFLIGHT_SET_POINT_H
